working bypass protection on eth_sendtransaction
