+++ src/lib/helpers/chrome/messageHandler.ts
@@ -2,3 +2,10 @@ export enum MessageType {
   ProceedAnyway = 'proceedAnyway',
   RunSimulation = 'runSimulation',
 }
+
+export const PortIdentifiers = {
+  WG_CONTENT_SCRIPT: 'wg-contentscript',
+  METAMASK_INPAGE: 'metamask-inpage',
+  METAMASK_CONTENT_SCRIPT: 'metamask-contentscript',
+  METAMASK_PROVIDER: 'metamask-provider',
+};
